requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.11
pyyaml==6.0.1
beautifulsoup4==4.12.3
python-dateutil==2.8.2
